requirements.txt
# Core deep learning
torch>=2.1.0
torchvision>=0.16.0

# Image augmentation
albumentations>=1.3.0

# Image I/O
Pillow>=10.0.0
numpy>=1.24.0

# Optional: experiment tracking
wandb>=0.16.0

# Optional: progress bars
tqdm>=4.66.0
